[Chore] Separate out optional dependency checks from vllm.utils (#27207)
