+++ tests/kernels/attention/test_deepgemm_attention.py
@@ -6,7 +6,7 @@ import pytest
 import torch
 
 from vllm.platforms import current_platform
-from vllm.utils import cdiv, has_deep_gemm
+from vllm.utils import cdiv
 from vllm.utils.deep_gemm import (
     _ceil_to_ue8m0,
     calc_diff,
@@ -15,6 +15,7 @@ from vllm.utils.deep_gemm import (
     get_num_sms,
     get_paged_mqa_logits_metadata,
 )
+from vllm.utils.import_utils import has_deep_gemm
 
 
 def kv_cache_cast_to_fp8(x: torch.Tensor) -> torch.Tensor:
