+++ tests/kernels/moe/parallel_utils.py
@@ -15,7 +15,7 @@ from torch.distributed import ProcessGroup
 from torch.multiprocessing import spawn  # pyright: ignore[reportPrivateImportUsage]
 from typing_extensions import ParamSpec
 
-from vllm.utils import has_deep_ep
+from vllm.utils.import_utils import has_deep_ep
 from vllm.utils.network_utils import get_open_port
 
 if has_deep_ep():
