+++ tests/kernels/moe/test_deepep_moe.py
@@ -21,7 +21,7 @@ from vllm.model_executor.layers.quantization.utils.fp8_utils import (
     per_token_group_quant_fp8,
 )
 from vllm.platforms import current_platform
-from vllm.utils import has_deep_ep
+from vllm.utils.import_utils import has_deep_ep
 
 from ...utils import multi_gpu_test
 from .parallel_utils import ProcessGroupInfo, parallel_launch
