+++ tests/models/quantization/test_fp8.py
@@ -12,7 +12,6 @@ from tests.quantization.utils import is_quant_method_supported
 from vllm.attention.utils.fa_utils import flash_attn_supports_fp8
 from vllm.platforms import current_platform
 from vllm.utils import STR_BACKEND_ENV_VAR
-
 from ..utils import check_logprobs_close
