+++ vllm/utils/deep_gemm.py
@@ -16,7 +16,8 @@ import torch
 import vllm.envs as envs
 from vllm.logger import logger
 from vllm.platforms import current_platform
-from vllm.utils import cdiv, has_deep_gemm
+from vllm.utils import cdiv
+from vllm.utils.import_utils import has_deep_gemm
 
 
 @functools.cache
